Split capability definitions into separate files.
