+++ kernel-ewasm/validator/src/serialization.rs
@@ -1,10 +1,8 @@
-use cap9_core::*;
 use crate::primitives::*;
 pub use core::fmt;
 use pwasm_std::vec::{Vec};
 
 use pwasm_std::String;
-use pwasm_std::types::*;
 
 
 /// Deserialization from serial i/o.
